Project- GenAI_Chat_Bot/requirements.txt
beautifulsoup4
langchain
nltk
numpy
openai
pandas
pypdf
python-dotenv
regex
requests
streamlit
touch